refactor: restructure project
- refactor OCXReader into lib (ocx) and executable (ocxreader), so lib functionality can be tested properly
