+++ ocx/include/ocx/ocx.h
@@ -1,0 +1,25 @@
+//
+// This file is part of OCXReader library
+// Copyright Carsten Zerbst (carsten.zerbst@groy-groy.de)
+//
+// This library is free software; you can redistribute it and/or modify it under
+// the terms of the GNU Lesser General Public License version 2.1 as published
+// by the Free Software Foundation.
+//
+
+// This is the main header file a user should include.
+
+#ifndef OCXREADERLIB_INCLUDE_OCX_OCX_H_
+#define OCXREADERLIB_INCLUDE_OCX_OCX_H_
+
+#include "ocx/ocx-caf-control-reader.h"
+#include "ocx/internal/ocx-context.h"
+#include "ocx/internal/ocx-coordinate-system-reader.h"
+#include "ocx/internal/ocx-curve-reader.h"
+#include "ocx/internal/ocx-helper.h"
+#include "ocx/internal/ocx-panel.h"
+#include "ocx/internal/ocx-panel-reader.h"
+#include "ocx/internal/ocx-reference-surfaces-reader.h"
+#include "ocx/internal/ocx-surface-reader.h"
+
+#endif //OCXREADERLIB_INCLUDE_OCX_OCX_H_
