+++ ocx/test/ocx-all-test.cc
@@ -1,0 +1,15 @@
+//
+// This file is part of OCXReader library
+// Copyright Carsten Zerbst (carsten.zerbst@groy-groy.de)
+//
+// This library is free software; you can redistribute it and/or modify it under
+// the terms of the GNU Lesser General Public License version 2.1 as published
+// by the Free Software Foundation.
+//
+
+// This file #includes all ocx-test implementation .cc files. The
+// purpose is to allow a user to build ocx-test by compiling this
+// file alone.
+
+// The following lines pull in the real ocx-test .cc files.
+#include "test/ocx-helper-test.cc"
